Most the changes on Spring.js t to render the users storage
Edit Spring.js to create a function that renders the users storage. An object that holds the values of each row. The function gets each value of every row and displays them on the screen. This will be useful when creating the buttons add and delete I could just put render list at the end to update the list