add remove button
make delete button functional. though there is room for improvement, as in the list of rows is never actually gone, say you delete 100 rows, all those rows still exist just have nothing in them to display. this could cause some space issues i am not to familiar with that kind of issue, as of now it works.