Make user interface editable
Make each value on the screen editable, besides the titles and header. Though very challenging and a small noticeable bug that is the variable name keeps returning a blank string and I could not figure it out. To turn this stone the function that used the variable name I just put that if the string is '' (empty) to just make it equal to what i needed it to be. at any rate it works and no user bugs are notice able as of now.